requirements.txt
requests>=2.31.0
edge-tts>=6.1.10
python-slugify>=8.0.0
ollama>=0.2.0
